added edit, delete and add methods to service
