--- src/app/list-todos.ts
@@ -1,10 +1,0 @@
-import { Todo } from './models/todo';
-
-export const TODOS: Todo[] = [
-  { id: 1, description: 'Make your bed', status: 'null', editable: false },
-  { id: 2, description: 'Make breakfast', status: 'null', editable: false },
-  { id: 3, description: 'Walk the dog', status: 'null', editable: false },
-  { id: 4, description: 'Buy bread', status: 'null', editable: false },
-  { id: 5, description: 'Buy eggs', status: 'null', editable: false },
-  { id: 6, description: 'Call landlord', status: 'null', editable: false },
-];
